Version 20240219.2
https://www.gurux.fi/release/guruxdlmscpp-202402191

+++ development/src/GXBigInteger.cpp
@@ -319,7 +319,7 @@ int CGXBigInteger::Add(CGXBigInteger& value)
                 tmp += value.Data[pos];
             }
             tmp += overflow;
-            Data[pos] = tmp;
+            Data[pos] = (uint32_t) tmp;
             overflow = tmp >> 32;
         }
         if (overflow != 0)
